+++ tests/step_defs/test_system.py
@@ -1,0 +1,10 @@
+"""System tests."""
+import testinfra_bdd
+from pytest_bdd import scenarios
+
+scenarios('../features/system.feature')
+
+
+# Ensure that the PyTest fixtures provided in testinfra-bdd are available to
+# your test suite.
+pytest_plugins = testinfra_bdd.PYTEST_MODULES