+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 
 ### New
 
+* Add Prometheus metrics. [Ben Dalling]
+
+* Add Trivy scan. [Ben Dalling]
+
+* Add a Docker Kafka broker as part of the test rig. [Ben Dalling]
+
 * Add the EnvironmentConfig class. [Ben Dalling]
 
 ### Other